Added "Discussion" section
I added 3 subreddits, 2 Discord servers, and 2 Matrix rooms that have a connection to the 'immutable' Linux desktop.  I am sure there are other communities out there that I am missing, but I listed the ones I can find.  I am not sure if this is necessary or if "Discussion" would be the appropriate title for this section as well.  However, here is my suggestion.  Nice page overall with very helpful resources.